+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## 1.0.4
 
-- Expose Readings.
+- Expose `readings`.
 - Breaking changes to match the `readings` getter:
   - Instead of `character.getMeanings(Language.english)`, use `character.meanings[Language.english]`.
   - Instead of `character.index.indexes`, use `character.indexes`.
 - Updated the embedded dictionary to version 2022-353 (2022-12-19).
+- Optimized initialization. Divides load time by 3.
 
 ## 1.0.3